requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
anthropic>=0.7.0
twilio==8.10.0
pydantic==2.4.2
python-multipart==0.0.6